+++ Backend/sacchon/src/main/java/gr/codehub/sacchon/model/BgMeasurement.java
@@ -27,7 +27,7 @@ public class BgMeasurement {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long BgMeasurementId ;
 
-    private int BgMeasurementData;
+    private Double BgMeasurementData;
     private LocalDate BgMeasurementDate;
 
     private LocalTime BgMeasurementTime;
